+++ test/chaining/select.test.js
@@ -1,7 +1,7 @@
 // test select (chaining of operations)
 var assert = require('assert'),
-    approx = require('../../../tools/approx.js'),
-    math = require('../../../index.js');
+    approx = require('../../tools/approx.js'),
+    math = require('../../index.js');
 
 describe('select', function() {
 
@@ -19,10 +19,10 @@ describe('select', function() {
 
   it('should chain operations with matrices', function() {
     assert.deepEqual(math.select(math.matrix([[1,2],[3,4]]))
-        .set(math.index(0,0), 8)
+        .subset(math.index(0,0), 8)
         .multiply(3).done(), math.matrix([[24, 6], [9, 12]]));
     assert.deepEqual(math.select([[1,2],[3,4]])
-        .set(math.index(0,0), 8)
+        .subset(math.index(0,0), 8)
         .multiply(3).done(), [[24, 6], [9, 12]]);
   });